Case 346: New customers will be prompted with a modal asking for zip code before seeing the plans and data page (IN PROGRESS)
Add modal with ZIP code form field if zipCode is not found in the session.
Do not allow user to close the modal without submitting a zipCode.

TO DO: (prerequisite: Case 573) The zip code validation should call the carrier of the device they are purchasing for to verify service with that carrier.