Update the phone.cfc to deal with Test server data changes.
A recent data change (in the last hour) broke all tests.  Update the code in Product.cfm to always launch the new devicebuilder modal whenever the 'Add to Cart' button was programmed to be displayed.

TODO: Someone will need to review this page for all of the logic that determines which button gets rendered in place of the 'Add to Cart' button.